Decided to go with Property changed Event
Changed NumberWordSwapper class from a static to an object so that I could add properties on it and capture when the properties changes.  Then added a Cached list that only holds the last 1000 records and has a OnProperty change event attached to it.  This way when the process hits 1000 records the user could still get the results of thta cache and work on those results before continuing with the rest.